+++ core/src/test/java/feign/TargetTest.java
@@ -17,9 +17,9 @@ package feign;
 
 import static feign.assertj.MockWebServerAssertions.assertThat;
 
-import com.squareup.okhttp.mockwebserver.MockResponse;
-import com.squareup.okhttp.mockwebserver.MockWebServer;
 import feign.Target.HardCodedTarget;
+import okhttp3.mockwebserver.MockResponse;
+import okhttp3.mockwebserver.MockWebServer;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -37,7 +37,7 @@ public class TargetTest {
   public void baseCaseQueryParamsArePercentEncoded() throws InterruptedException {
     server.enqueue(new MockResponse());
 
-    String baseUrl = server.getUrl("/default").toString();
+    String baseUrl = server.url("/default").toString();
 
     Feign.builder().target(TestQuery.class, baseUrl).get("slash/foo", "slash/bar");
 
@@ -52,7 +52,7 @@ public class TargetTest {
   public void targetCanCreateCustomRequest() throws InterruptedException {
     server.enqueue(new MockResponse());
 
-    String baseUrl = server.getUrl("/default").toString();
+    String baseUrl = server.url("/default").toString();
     Target<TestQuery> custom =
         new HardCodedTarget<TestQuery>(TestQuery.class, baseUrl) {