+++ core/src/test/java/feign/assertj/RecordedRequestAssert.java
@@ -18,8 +18,6 @@ package feign.assertj;
 import static org.assertj.core.data.MapEntry.entry;
 import static org.assertj.core.error.ShouldNotContain.shouldNotContain;
 
-import com.squareup.okhttp.Headers;
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
 import feign.Util;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -30,6 +28,8 @@ import java.util.Map;
 import java.util.Set;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.InflaterInputStream;
+import okhttp3.Headers;
+import okhttp3.mockwebserver.RecordedRequest;
 import org.assertj.core.api.AbstractAssert;
 import org.assertj.core.data.MapEntry;
 import org.assertj.core.internal.ByteArrays;